requirements.txt
Flask==1.0.2
gunicorn==19.8.1
line-bot-sdk==1.7.2
numpy==1.13.3
python-dotenv==0.8.2